Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    double a,b;
+    float a,b;
     char c;
     scanf("%f %f %c",&a,&b,&c);
     if (c=='+') {
